adding error log when error boundary event is thrown
