+++ src/main/java/de/medizininformatik_initiative/process/report/service/HandleError.java
@@ -1,35 +1,46 @@
 package de.medizininformatik_initiative.process.report.service;
 
+import java.util.Objects;
+
 import org.hl7.fhir.r4.model.Task;
+import org.springframework.beans.factory.InitializingBean;
 
 import de.medizininformatik_initiative.process.report.ConstantsReport;
+import de.medizininformatik_initiative.process.report.util.ReportStatusGenerator;
 import de.medizininformatik_initiative.processes.common.util.ConstantsBase;
 import dev.dsf.bpe.v2.ProcessPluginApi;
 import dev.dsf.bpe.v2.activity.ServiceTask;
 import dev.dsf.bpe.v2.client.dsf.DelayStrategy;
 import dev.dsf.bpe.v2.variables.Variables;
 
-public class HandleError implements ServiceTask
+public class HandleError implements ServiceTask, InitializingBean
 {
+	private final ReportStatusGenerator statusGenerator;
 	private final boolean hrpEmailEnabled;
 
-	public HandleError(boolean hrpEmailEnabled)
+	public HandleError(ReportStatusGenerator statusGenerator, boolean hrpEmailEnabled)
 	{
+		this.statusGenerator = statusGenerator;
 		this.hrpEmailEnabled = hrpEmailEnabled;
+	}
+
+	@Override
+	public void afterPropertiesSet() throws Exception
+	{
+		Objects.requireNonNull(statusGenerator, "reportStatusGenerator");
 	}
 
 	@Override
 	public void execute(ProcessPluginApi api, Variables variables)
 	{
 		Task task = variables.getStartTask();
+		String errorCode = variables.getString(ConstantsReport.BPMN_EXECUTION_VARIABLE_REPORT_RECEIVE_ERROR);
+		String errorMessage = variables.getString(ConstantsReport.BPMN_EXECUTION_VARIABLE_REPORT_RECEIVE_ERROR_MESSAGE);
+
 		if (hrpEmailEnabled)
 			sendMail(api, variables, task);
 
-		if (Task.TaskStatus.FAILED.equals(task.getStatus()))
-		{
-			api.getDsfClientProvider().getLocal().withRetry(ConstantsBase.DSF_CLIENT_RETRY_6_TIMES,
-					DelayStrategy.constant(ConstantsBase.DSF_CLIENT_RETRY_INTERVAL_5MIN)).update(task);
-		}
+		failAndAddOutputTask(api, task, errorCode, errorMessage, variables);
 	}
 
 	private void sendMail(ProcessPluginApi api, Variables variables, Task task)
@@ -47,5 +58,17 @@ public class HandleError implements ServiceTask
 				+ (error == null ? "none" : error);
 
 		api.getMailService().send(subject, message);
+	}
+
+	private void failAndAddOutputTask(ProcessPluginApi api, Task task, String errorCode, String errorMessage,
+			Variables variables)
+	{
+		task.setStatus(Task.TaskStatus.FAILED);
+		task.addOutput(statusGenerator.createReportStatusOutput(api.getProcessPluginDefinition().getResourceVersion(),
+				errorCode, errorMessage));
+		variables.updateTask(task);
+
+		api.getDsfClientProvider().getLocal().withRetry(ConstantsBase.DSF_CLIENT_RETRY_6_TIMES,
+				DelayStrategy.constant(ConstantsBase.DSF_CLIENT_RETRY_INTERVAL_5MIN)).update(task);
 	}
 }
